pyproject.toml
[tool.poetry]
name = "ai-expert-assistant-backend"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langserve = {extras = ["all"], version = "^0.3.0"}
openai = "^1.54.4"
langgraph = "^0.2.48"
langchain-openai = "^0.2.8"
python-dotenv = "^1.0.1"
azure-identity = "^1.19.0"
python-decouple = "^3.8"
pytz = "^2024.2"
apscheduler = "^3.10.4"
cachetools = "^5.5.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
psycopg-pool = "^3.2.4"
langgraph-checkpoint-postgres = "^2.0.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
tzlocal = "^5.2"
azure-monitor-opentelemetry-exporter = "^1.0.0b32"
openinference-instrumentation-langchain = "^0.1.29"
opentelemetry-sdk = "^1.28.2"
opentelemetry-exporter-otlp = "^1.28.2"
azure-monitor-opentelemetry = "^1.6.4"
python-docx = "^1.1.2"
python-multipart = "^0.0.19"
azure-ai-documentintelligence = "^1.0.0b4"
pycryptodome = "^3.21.0"
python-dateutil = "^2.9.0.post0"
azure-cosmos="^4.9.0"
aiohttp="^3.11.18"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
